Local Storage randomness
Add a benchmark for LocalStorage that writes random bytes and
has some other tweaks as well. This is less to test performance
(LocalStorage write speed is unrelated to the backend) and more
to have a way to trigger Chrome to write lots of LocalStorage
changes. With diagnostic tools like `iotop`, we can see
how much Chrome is writing to disk for LocalStorage.